requirements.txt
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
torchmetrics
numpy
tqdm
matplotlib
pandas
wandb
hydra_core
speechbrain~=0.5.12
torch_audiomentations
omegaconf==2.3.0

black
isort
pre-commit
